+++ lib-SLProject/source/CV/OrbSlam/Converter.cpp
@@ -65,7 +65,7 @@ cv::Mat Converter::toCvMat(const Eigen::Matrix<double,4,4> &m)
     cv::Mat cvMat(4,4,CV_32F);
     for(int i=0;i<4;i++)
         for(int j=0; j<4; j++)
-            cvMat.at<float>(i,j)=m(i,j);
+            cvMat.at<float>(i,j)=(float)m(i,j);
 
     return cvMat.clone();
 }
@@ -75,7 +75,7 @@ cv::Mat Converter::toCvMat(const Eigen::Matrix3d &m)
     cv::Mat cvMat(3,3,CV_32F);
     for(int i=0;i<3;i++)
         for(int j=0; j<3; j++)
-            cvMat.at<float>(i,j)=m(i,j);
+            cvMat.at<float>(i,j)=(float)m(i,j);
 
     return cvMat.clone();
 }
@@ -84,7 +84,7 @@ cv::Mat Converter::toCvMat(const Eigen::Matrix<double,3,1> &m)
 {
     cv::Mat cvMat(3,1,CV_32F);
     for(int i=0;i<3;i++)
-            cvMat.at<float>(i)=m(i);
+            cvMat.at<float>(i)=(float)m(i);
 
     return cvMat.clone();
 }
